Add base Controller class
Controller class is used as a basis to extend all other controllers.
At this point is has one method: 'view' that loads a view based on
received parameter and passes data to that view. Next stage of
development of Controller is to add a method to load models, so that
controllers can work with database.