Refactor constructor method of Core class
This refactors the constructor method of the Core class to better parse
URL. See the comment to the constructor method for more details.

This also adds some additional comments.

Any hyphens in URL are now
removed during URL parsing since the methods are called based on URL,
and and method names cannot contain hyphens in PHP.