Change the algorithm for deleting checkboxes
Attempts to avoid the problem of autotest failing to delete products
from main page due to not actually clicking on checkboxes but rather
changing the '.checked' status of all checkboxes at once.

The previous algorithm disabled 'MASS DELETE' button of the main page
by default, and put listeners on all checkboxes: after a checkbox was
clicked, the listener added the checkbox's value to an array. The value
was removed from the array when a checkbox was unchecked. If the array
was not empty, the button was enabled, and was disabled when array was
empty again.

The previous algorithm made it possible to have disabled mass delete
button that only became functional when any checkbox was selected.

In the current algorithm the button is always enabled. When clicked, it
finds all checked checkboxes and sends their values to the endpoint
that deletes products matching their values.